feat(hub-common): fix cyclical module dependency introduced on this PR
affects: @esri/hub-common

ISSUES CLOSED: 12495
